for #2084, move package of PredicateInRightValueSegment
